Completed code

+++ get_next_line.h
@@ -12,7 +12,7 @@
 //#include <stdlib.h> and <unistd.h> for *malloc(), free(), & read()*.
 
 char	*ft_substr(char const *s, unsigned int start, size_t len);
-int		ft_strlen(const char *str);
+size_t	ft_strlen(const char *str);
 char	*ft_strchr(const char *s, int i);
 char	*ft_strdup(const char *s);
 char	*ft_strjoin(char const *s1, char const *s2);
